pyproject.toml
[tool.poetry]
name = "kubernetes-django"
version = "0.1.0"
description = ""
authors = ["Oussama elamraoui"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
build = "^1.2.2.post1"
cachecontrol = "^0.14.1"
certifi = "^2024.12.14"
cffi = "^1.17.1"
charset-normalizer = "^3.4.1"
cleo = "^2.1.0"
crashtest = "^0.4.1"
cryptography = "^44.0.0"
distlib = "^0.3.9"
django = "^5.1.4"
djangorestframework = "^3.15.2"
fastjsonschema = "^2.21.1"
filelock = "^3.16.1"
idna = "^3.10"
installer = "^0.7.0"
jaraco-classes = "^3.4.0"
jeepney = "^0.8.0"
more-itertools = "^10.5.0"
msgpack = "^1.1.0"
packaging = "^24.2"
pexpect = "^4.9.0"
pkginfo = "^1.12.0"
platformdirs = "^4.3.6"
poetry = "^1.8.5"
poetry-core = "^1.9.1"
poetry-plugin-export = "^1.8.0"
ptyprocess = "^0.7.0"
pycparser = "^2.22"
pyproject-hooks = "^1.2.0"
python-dotenv = "^1.0.1"
rapidfuzz = "^3.11.0"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
secretstorage = "^3.3.3"
shellingham = "^1.5.4"
sqlparse = "^0.5.3"
tomlkit = "^0.13.2"
trove-classifiers = "^2024.10.21.16"
urllib3 = "^2.3.0"
virtualenv = "^20.28.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
flake8 = "^7.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
